+++ packages/kysely/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @authhero/kysely-adapter
 
+## 8.1.0
+
+### Minor Changes
+
+- Update entities for sessions and refresh tokens
+
+### Patch Changes
+
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.45.0
+
 ## 8.0.0
 
 ### Minor Changes